Track guard changes, fix bugs, refactor code. Also update the svn:external to the new URL.
git-svn-id: https://svn.torproject.org/svn/torflow/trunk@21402 55e972cd-5a19-0410-ae62-a4d7a52db4cd

+++ CircuitAnalysis/BuildTimes/CBT-Test/run_test.sh
@@ -3,7 +3,7 @@
 # 1. Fire up a Tor
 # 2. For every 5% percentile, loop N times:
 # 3.   Remove state file + hup tor.
-# 4.   Run test with 3 guards from a percentile for c>10 circuits:
+# 4.   Run test with 3 guards from a percentile for c>50 circuits:
 # 5.     Record MIN_CIRCS=c/2 when past c/2 circuits did not change
 #        timeout more than +/-1.
 # 6.       Copy state file to saved location w/ guards and c/2
@@ -34,48 +34,55 @@
 #       though..
 # C. SETCONF EntryNodes/StrictEntryNodes
 
-TOR_DIR=../../tor.git/src/or
+TOR_DIR=../../../../tor.git/src/or
 TOR_DATA=./tor-data/
 
-kill `cat $TOR_DATA/tor.pid`
-if [ $? -eq 0 ]
-then
- sleep 10
+if [ -f $TOR_DATA/tor.pid ]; then
+  kill `cat $TOR_DATA/tor.pid`
+  if [ $? -eq 0 ]
+  then
+   sleep 10
+  fi
 fi
 
-$TOR_DIR/tor -f $TOR_DATA/torrc
-
-# XXX: Handle no guards left case
-for p in 65 0 5 10 15 20 25 30 35 40 45 50 55 60
+for p in 0 10 20 30 40 50 60 70 80 90
 do
-  rm $TOR_DATA/state
-  kill -HUP `cat $TOR_DATA/tor.pid`
   N=0
   while [ $N -lt 10 ]
   do
-    mkdir -p results/$p/$N/min
-    mkdir -p results/$p/$N/full
-    echo ./cbt-test.py -m -P 5 -p $p -o results/$p/$N/min/result -b results/$p/$N/min/buildtimes
-    # XXX: check retval
-    cp $TOR_DATA/state results/$p/$N/min/state
-    echo ./cbt-test.py -f -P 5 -p $p -o results/$p/$N/full/result -b results/$p/$N/full/buildtimes 
-    cp $TOR_DATA/state results/$p/$N/full/state
+    if [ -f $TOR_DATA/tor.pid ]; then
+      kill `cat $TOR_DATA/tor.pid`
+      wait `cat $TOR_DATA/tor.pid`
+    fi
+    rm $TOR_DATA/state
+    $TOR_DIR/tor -f $TOR_DATA/torrc &
+    sleep 10
+    mkdir -p results/$p/$N
+    ./cbttest.py -p $p -o results/$p/$N || exit
     N=`expr $N + 1`
   done
 done
+
+exit
 
 for p in `ls -1 results`
 do
   for n in `ls -1 results/$p`
   do
-    for t in `ls -1 results/$p/$n`
+    for state in `ls -1 results/$p/$n/state.*`
     do
       M=0
       while [ $M -lt 3 ]
       do
-        cp results/$p/$n/$t $TOR_DATA/state
-        kill -HUP `cat $TOR_DATA/tor.pid`
-        echo ./cbt-test.py -r -o results/$p/$n/$t/result -b results/$p/$N/buildtimes.redo.$M
+        if [ -f $TOR_DATA/tor.pid ]; then
+          kill `cat $TOR_DATA/tor.pid`
+          wait `cat $TOR_DATA/tor.pid`
+        fi
+        cp $state $TOR_DATA/state
+        $TOR_DIR/tor -f $TOR_DATA/torrc &
+        sleep 10
+        # XXX: M times?? need diff result files..
+        ./cbt-test.py -r -o results/$p/$n || exit
         M=`expr $N + 1`
       done
     done
